pyproject.toml
[project]
name = "fvg-detector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.3.3",
    "python-dotenv>=1.0.0",
    "yfinance>=0.2.50",
]
